requirements.txt
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.3.0
python-docx>=1.1.0
google-generativeai>=0.3.2
edge-tts>=6.1.0
pydub>=0.25.1
sentence-transformers>=2.7.0
chromadb>=0.5.5
streamlit>=1.28.0
